+++ packages/terra-search-field/tests/nightwatch/SearchFieldTests.jsx
@@ -9,8 +9,9 @@ const SearchFieldTests = () => (
       <li><Link to="/tests/search-field-tests/default">SearchField - Default</Link></li>
       <li><Link to="/tests/search-field-tests/placeholder">SearchField - Placeholder</Link></li>
       <li><Link to="/tests/search-field-tests/callback">SearchField - Callback for Valid and Invalid Entry</Link></li>
-      <li><Link to="/tests/search-field-tests/delayed">SearchField - 1000ms Delayed </Link></li>
-      <li><Link to="/tests/search-field-tests/minimum-length">SearchField - 5 Character Minimum Length </Link></li>
+      <li><Link to="/tests/search-field-tests/delayed">SearchField - 1000ms Delayed</Link></li>
+      <li><Link to="/tests/search-field-tests/minimum-length">SearchField - 5 Character Minimum Length</Link></li>
+      <li><Link to="/tests/search-field-tests/stretched">SearchField - Stretched</Link></li>
     </ul>
   </div>
 );