+++ packages/terra-search-field/tests/nightwatch/search-field-spec.js
@@ -74,4 +74,15 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
       });
     });
   },
+
+  'Displays a search field that is strecthed to fill its container': (browser) => {
+    browser
+      .url(`${browser.launchUrl}/#/tests/search-field-tests/stretched`)
+      .assert.elementPresent('#searchfield')
+      .assert.elementPresent('#searchfield button')
+      .assert.attributeEquals('#searchfield input', 'placeholder', '');
+
+    browser.expect.element('#searchfield').to.have.attribute('class').which.contains('searchfield');
+    browser.expect.element('#searchfield').to.have.attribute('class').which.contains('stretch');
+  },
 });